requirements.txt
APScheduler==3.11.0
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
Django==4.2.18
django-scheduler==0.10.1
filelock==3.18.0
fsspec==2025.5.1
hf-xet==1.1.2
huggingface-hub==0.32.2
icalendar==6.1.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
load-dotenv==0.1.0
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pyTelegramBotAPI==4.26.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
schedule==1.2.2
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==4.1.0
six==1.17.0
sqlparse==0.5.3
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.52.3
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.3.1
urllib3==2.3.0